requirements.txt
streamlit
ultralytics
opencv-python-headless
numpy
Pillow
torch